méthode pour la création d'article

+++ app/Http/Controllers/ArticlesController.php
@@ -4,10 +4,13 @@ namespace App\Http\Controllers;
 
 use App\Models\Article;
 use App\Models\User;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
 
-class ArticlesController extends Controller
+
+class ArticlesController extends Controller 
 {
+    use AuthorizesRequests;
     public function index()
     {
         $articles = Article::with('user')->orderBy('created_at', 'DESC')->get();
@@ -50,13 +53,17 @@ class ArticlesController extends Controller
 
     public function create()
     {
+        $this->authorize('create', arguments: Article::class);
         return view('articles.create',);
     }
 
     public function store(Request $request)
     {
+        $this->authorize('create', arguments: Article::class);
         // vérification des permissions plus tard
-        $user = User::find(1);
+        // $user = User::find(1);
+        $user = auth()->user(); //On récupère l'utilisateur connecté 
+
         $request['user_id'] = $user->id;
 
         $validatedData = $request->validate([
@@ -73,17 +80,20 @@ class ArticlesController extends Controller
 
     public function edit(Article $article)
     {
+        $this->authorize('update', $article);
         return view('articles.edit', compact('article'));
     }
 
     public function update(Request $request, Article $article)
     {
+        $this->authorize('update', $article);
         // dd($article, $request->all());
         $article->update($request->all());
         return redirect('/articles')->with(['success_message' => 'L\'article a été modifiée !']);
     }
 
     public function delete(Article $article){
+        $this->authorize('delete', $article);
         // vérification des permissions plus tard
         $article->delete();
         return redirect('/articles')->with(['success_message' => 'L\'article a été supprimée !']);
